Move authorize into its own script
